The scheduler deamon module is starting up by default, which results in all nodes in the cluster are running the scheduler and will start jobs.
The scheduler get's only started when the system property project.config is specifying a path at which the project configuration properties are defining if the scheduler should be active. If the system property project.config is not set a lookup on the classpath is done for the resource /project.properties

A future improvement for the scheduler could be to use the repository and workflow for scheduling, for details have a look at the broken links addon.